+++ modular_core/lib/src/di/bind_context.dart
@@ -19,14 +19,15 @@ abstract class BindContextImpl implements BindContext {
   @internal
   final Set<String> tags = {};
   final _singletonBinds = <Type, SingletonBind>{};
-  List<SingletonBind> get instanciatedSingletons =>
-      _singletonBinds.values.toList();
+
+  @override
+  List<SingletonBind> get instanciatedSingletons => _singletonBinds.values.toList();
 
   @visibleForTesting
   List<BindContract> getProcessBinds() => _binds;
 
   void changeBinds(List<BindContract> newBinds) {
-    _binds.clear();
+    _binds.removeWhere((element) => !element.alwaysSerialized);
     _binds.addAll(newBinds);
   }
 
@@ -50,9 +51,7 @@ abstract class BindContextImpl implements BindContext {
       return bindValue;
     }
 
-    var bind = _binds.firstWhere(
-        (b) => b.factoryFunction is T Function(Injector),
-        orElse: () => BindEmpty());
+    var bind = _binds.firstWhere((b) => b.factoryFunction is T Function(Injector), orElse: () => BindEmpty());
     if (bind is BindEmpty) {
       return null;
     }
@@ -106,27 +105,22 @@ abstract class BindContextImpl implements BindContext {
     _mutableValue.isReadyFlag = true;
     final asyncBindList = _binds.whereType<AsyncBindContract>().toList();
     for (var bind in asyncBindList) {
-      final resolvedBind = await bind.convertToAsyncBind();
+      final resolvedBind = await bind.convertToBind();
       _binds.insert(0, resolvedBind);
     }
   }
 
   @mustCallSuper
-  void instantiateSingletonBinds(
-      List<SingletonBind> singletons, Injector injector) {
-    final filteredList =
-        _binds.where((bind) => !bind.isLazy && !_containBind(singletons, bind));
+  void instantiateSingletonBinds(List<SingletonBind> singletons, Injector injector) {
+    final filteredList = _binds.where((bind) => !bind.isLazy && !_containBind(singletons, bind));
     for (final bindElement in filteredList) {
       var b = bindElement.factoryFunction(injector);
-      _singletonBinds[b.runtimeType] =
-          SingletonBind(value: b, bind: bindElement);
+      _singletonBinds[b.runtimeType] = SingletonBind(value: b, bind: bindElement);
     }
   }
 
   bool _containBind(List<SingletonBind> singletons, BindContract bind) {
-    return singletons.indexWhere((element) =>
-            element.bind.factoryFunction == bind.factoryFunction) !=
-        -1;
+    return singletons.indexWhere((element) => element.bind.factoryFunction == bind.factoryFunction) != -1;
   }
 
   Type _getInjectType<B>() {
@@ -134,9 +128,7 @@ abstract class BindContextImpl implements BindContext {
 
     for (var singleton in _singletonBinds.values) {
       if (singleton.value is B) {
-        foundType = _singletonBinds.entries
-            .firstWhere((map) => map.value.value == singleton.value)
-            .key;
+        foundType = _singletonBinds.entries.firstWhere((map) => map.value.value == singleton.value).key;
         break;
       }
     }