+++ modular_core/lib/src/di/reassemble_mixin.dart
@@ -1,0 +1,3 @@
+mixin ReassembleMixin {
+  void reassemble();
+}