+++ modular_core/lib/src/route/route_context.dart
@@ -3,10 +3,12 @@ import 'package:modular_interfaces/modular_interfaces.dart';
 
 import '../di/bind_context.dart';
 
-abstract class RouteContextImpl extends BindContextImpl
-    implements RouteContext {
+abstract class RouteContextImpl extends BindContextImpl implements RouteContext {
   @override
   List<ModularRoute> get routes => const [];
+
+  @override
+  final List<RouteContext> modules = [];
 
   @visibleForTesting
   List<ModularKey> orderRouteKeys(Iterable<ModularKey> keys) {
@@ -19,8 +21,7 @@ abstract class RouteContextImpl extends BindContextImpl
       if (preview.name.contains('**')) {
         if (!actual.name.contains('**')) {
           return 1;
-        } else if (actual.name.split('/').length >
-            preview.name.split('/').length) {
+        } else if (actual.name.split('/').length > preview.name.split('/').length) {
           return 1;
         }
       }
@@ -32,6 +33,8 @@ abstract class RouteContextImpl extends BindContextImpl
 
   @override
   Map<ModularKey, ModularRoute> init() {
+    modules.clear();
+    modules.add(this);
     final _routeMap = <ModularKey, ModularRoute>{};
     for (var route in routes) {
       _routeMap.addAll(assembleRoute(route));
@@ -62,10 +65,9 @@ abstract class RouteContextImpl extends BindContextImpl
   Map<ModularKey, ModularRoute> addModule(ModularRoute route) {
     final Map<ModularKey, ModularRoute> map = {};
     final module = route.context!;
+    modules.add(module);
     for (var child in module.routes) {
-      child = child.copyWith(
-          bindContextEntries: {module.runtimeType: module},
-          parent: route.parent);
+      child = child.copyWith(bindContextEntries: {module.runtimeType: module}, parent: route.parent);
       child = copy(route, child);
       map.addAll(assembleRoute(child));
     }