+++ modular_core/test/src/di/bind_context_test.dart
@@ -102,8 +102,7 @@ void main() {
   });
 
   test('instantiateSingletonBinds', () {
-    instance.instantiateSingletonBinds(
-        [SingletonBind(bind: _Bind((i) => 0.0), value: 0.0)], injector);
+    instance.instantiateSingletonBinds([SingletonBind(bind: _Bind((i) => 0.0), value: 0.0)], injector);
     expect(instance.instanciatedSingletons.length, 1);
   });
 }
@@ -150,13 +149,11 @@ abstract class IRepository {}
 
 class Repository extends IRepository {}
 
-class AsyncBind<T extends Object> extends _Bind<Future<T>>
-    implements AsyncBindContract<T> {
+class AsyncBind<T extends Object> extends _Bind<Future<T>> implements AsyncBindContract<T> {
   @override
   final Future<T> Function(Injector i) asyncInject;
 
-  AsyncBind(this.asyncInject, {bool export = false})
-      : super(asyncInject, export: export);
+  AsyncBind(this.asyncInject, {bool export = false}) : super(asyncInject, export: export);
 
   @override
   Future<T> resolveAsyncBind() async {
@@ -165,7 +162,7 @@ class AsyncBind<T extends Object> extends _Bind<Future<T>>
   }
 
   @override
-  Future<BindContract<T>> convertToAsyncBind() async {
+  Future<BindContract<T>> convertToBind() async {
     final bindValue = await resolveAsyncBind();
     return _Bind<T>((i) => bindValue, export: export);
   }