+++ flutter_modular/CHANGELOG.md
@@ -1,3 +1,6 @@
+## [4.1.0] - 2021-10-01
+* Modular is now HOT-RELOAD friendly!
+
 ## [4.0.1+1] - 2021-09-22
 * Fixed pushNamed.
 * Fixed bug that allowed access to parameters and arguments in other modules.